Move to wrapped lowercase and unwrapped lowercase.
